requirements.txt
snowflake-core==0.9.0
snowflake-connector-python>=2.8.0
snowflake-snowpark-python<2.0.0
streamlit>=1.28.0
streamlit-lottie
pandas>=1.3.0
